+++ src/Paramore.Brighter.MessagingGateway.RMQ/RmqMessageProducer.cs
@@ -67,7 +67,7 @@ namespace Paramore.Brighter.MessagingGateway.RMQ
         {
             lock (_lock)
             {
-                _logger.Value.DebugFormat("RmqMessageProducer: Preparing  to sending message via exchange {0}", Connection.Exchange.Name);
+                _logger.Value.DebugFormat("RmqMessageProducer: Preparing  to send message via exchange {0}", Connection.Exchange.Name);
                 EnsureChannel();
                 var rmqMessagePublisher = new RmqMessagePublisher(Channel, Connection.Exchange.Name);
                 _logger.Value.DebugFormat("RmqMessageProducer: Publishing message to exchange {0} on connection {1} with a delay of {5} and topic {2} and id {3} and body: {4}", Connection.Exchange.Name, Connection.AmpqUri.GetSanitizedUri(), message.Header.Topic, message.Id, message.Body.Value, delayMilliseconds);