+++ release_notes.md
@@ -7,6 +7,8 @@ This section lists features in master, available by [AppVeyor](https://ci.appvey
 
 ## Master ##
 
+  - Added Support for a Redis transport
+
 ## Release 7.1.0 ##
   - Fixes issue with high CPU when failing to connect to RabbitMQ.
   - Fixes missing High Availability setting, had to make changes to IAmAChannelFactory.